requirements.txt
PySide6==6.7.2
paramiko==3.4.0
PyYAML==6.0.2
Jinja2==3.1.6
colorama==0.4.6